requirements.txt
discord.py==2.5.2
python-dotenv==1.0.0
feedparser==6.0.10
Flask==3.1.1
Thread==2.0.5